Fix : 이벤트 추가&수정 subView 분리
